fix(checker): split tuple alias display helper
